+++ js/api.js
@@ -2,7 +2,7 @@ const BASE_URL = 'https://31.javascript.htmlacademy.pro/kekstagram';
 
 const Route = {
   GET_DATA: '/data',
-  SEND_DATA:'/1',
+  SEND_DATA:'/',
 };
 
 const Method = {
